Can now edit todos, createTodo renamed to saveTodo, space added at eof's.
Changed all instances of createTodo to saveTodo since it is being used in the update function and makes more sense to be called that. The way the todos are being updated right now, doesn't seem to be the best way since handleSubmit is calling two actions when it should only be calling one (updateTodo) but the way the array is set up I can only either call text or id not the entire object itself.